requirements.txt
discord.py
beautifulsoup4
lxml
scipy
scikit-image
requests
pytest
isort
black
